projects/deployment/activelog-deploy/lambda/requirements.txt
boto3 
requests 
openai 
python-dotenv 
